requirements.txt
flask
faiss-cpu
numpy
python-dotenv
langchain-openai
langchain-core
prometheus_client
openai